+++ CHANGELOG.md
@@ -8,7 +8,7 @@ All notable changes to this project will be documented in this file.
   - `getOption(string)` ([#75](https://github.com/rcjsuen/dockerfile-ast/issues/75))
   - `hasOptions()` ([#75](https://github.com/rcjsuen/dockerfile-ast/issues/75))
 - `FlagOption` ([#75](https://github.com/rcjsuen/dockerfile-ast/issues/75))
-- `Run` ([#75](https://github.com/rcjsuen/dockerfile-ast/issues/76))
+- `Run` ([#76](https://github.com/rcjsuen/dockerfile-ast/issues/76))
   - `RUN` instructions will now of type `Run` instead of `JSONInstruction`
   - `Run` extends `JSONInstruction` so this is not be a breaking change