+++ CHANGELOG.md
@@ -3,7 +3,7 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 ### Fixed
-- correct range of multiline arguments that end with an escaped quote ([#78](https://github.com/rcjsuen/dockerfile-ast/issues/78))
+- correct range of multiline arguments that end with a character preceded by an escape character ([#78](https://github.com/rcjsuen/dockerfile-ast/issues/78))
 
 ## [0.0.25] - 2020-04-01
 ### Added